# pow: correctly compute exactness