Power calculations of x**y where y is an integer will now correctly
be flagged as exact when possible. This avoids lots of work and is
thus much faster for these operations.

Remove some unneeded precision changing in integerPower. It
was leftover from the inf code which did bad things with 1/x
calculations. This was causing some overflow problems that shouldn't
have been present.

This change reduces `go test` time from 2.5s to 0.9s.

```
name          old time/op  new time/op  delta
GDA/power-12   666ms ±16%   387ms ± 6%  -41.86%  (p=0.008 n=5+5)
```

---

First commit is from another PR, but has a fix this needs.

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/apd/30)
<!-- Reviewable:end -->