@eisenstatdavid 2017-03-08
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/> pending rebase, but I'm wondering if it would make sense to use integerPower on the integer part and do the log/mul/exp calculation on the fractional part and multiply. That implementation would be obviously monotone, which is worth something, and it might be faster (though I really have no idea; if it isn't, does this PR slow some cases with large exponents down?)

---

Reviewed 2 of 2 files at r1.
Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---

*[context.go, line 864 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/30#-KeiY99FOcwHdMhaGCYS:-KeiY99FOcwHdMhaGCYT:b-yzfqfd) ([raw file](https://github.com/cockroachdb/apd/blob/7884ea1a6ef6ecf37ed0bc57d490b61d0bec026c/context.go#L864)):*
> ```Go
> 	}
> 	// 1 ** x == 1
> 	if x.Cmp(decimalOne) == 0 {
> ```

This case isn't handled specially by the new code, is it?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/30#-:-KeiZcqxF1hI9wVKHzFG:b-s7m4h)*
<!-- Sent from Reviewable.io -->