@mjibson 2017-03-08
I can't figure out how to split up and recombine the integer and fractional parts to test if this is faster. Is there an identity I'm missing?

---

Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---

*[context.go, line 864 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/30#-KeiY99FOcwHdMhaGCYS:-KejCzc0S2EPozha_mRK:b-c42e9w) ([raw file](https://github.com/cockroachdb/apd/blob/7884ea1a6ef6ecf37ed0bc57d490b61d0bec026c/context.go#L864)):*
<details><summary><i>Previously, eisenstatdavid (David Eisenstat) wrote…</i></summary><blockquote>

This case isn't handled specially by the new code, is it?
</blockquote></details>

No, but it doesn't need to be. It was handled specially as a performance improvement, but as a side effect it incorrectly flagged exactness. Pow is fast enough now that we don't need to special case it.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/30#-:-Kej_PO8RV-l0rJDuEQ-:br08oit)*
<!-- Sent from Reviewable.io -->