@eisenstatdavid 2017-03-08
pow(x, y) = pow(x, integ(y)) * pow(x, frac(y))?

---

Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/30#-:-Kejb6Vhm7uJn_cXZKrr:b8lfa3c)*
<!-- Sent from Reviewable.io -->