@mjibson 2017-03-08
I don't think this PR slows down integer cases with large exponents. If the exponents are very large, then they will quickly overflow anyway, so I've been having difficulty constructing an interesting test case.

---

Review status: 0 of 2 files reviewed at latest revision, 2 unresolved discussions.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/30#-:-KejayQQCc98rt04esPB:bd95mn4)*
<!-- Sent from Reviewable.io -->