@mjibson 2017-03-08
Oh duh. I was thinking about separating x into integ and frac parts. Not thinking clearly.

---

Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/30#-:-KejbOTDHvN5-ha4fMot:b-fxmzon)*
<!-- Sent from Reviewable.io -->