@eisenstatdavid 2017-03-08
Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---

*[context.go, line 864 at r1](https://reviewable.io:443/reviews/cockroachdb/apd/30#-KeiY99FOcwHdMhaGCYS:-KejcMQLEOLCzZpD9edP:bksy8p7) ([raw file](https://github.com/cockroachdb/apd/blob/7884ea1a6ef6ecf37ed0bc57d490b61d0bec026c/context.go#L864)):*
<details><summary><i>Previously, mjibson (Matt Jibson) wrote…</i></summary><blockquote>

No, but it doesn't need to be. It was handled specially as a performance improvement, but as a side effect it incorrectly flagged exactness. Pow is fast enough now that we don't need to special case it.
</blockquote></details>

Weird omission from the spec. OK.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/30)*
<!-- Sent from Reviewable.io -->