@mjibson 2017-03-08
You were right. Separating the integer and fractional calculations out yielded another 10% speed up. Added a second commit with this. RFAL.

---

Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/30#-:-KejeLO8uF1XXIWmqrfQ:b-grjfxw)*
<!-- Sent from Reviewable.io -->