@eisenstatdavid 2017-03-08
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 2 files reviewed at latest revision, 1 unresolved discussion.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/apd/30#-:-KejhqTG7i7A5iqva-ih:bnfp4nl)*
<!-- Sent from Reviewable.io -->